+++ app.js
@@ -5,7 +5,6 @@ const locations = require("./controllers/locations.controller");
 const machines = require("./controllers/machines.controller");
 const people = require("./controllers/persons.controller");
 const plans = require("./controllers/plans.controller");
-const peopleByLocation = require("./controllers/peopleByLocation.controller");
 
 const app = express();
 
@@ -14,7 +13,6 @@ app.use("/machines", machines);
 app.use("/persons", people);
 app.use("/plans", plans);
 app.use("/special-events", specialEvents);
-app.use("/locations/people", peopleByLocation);
 
 app.get("/", (req, res) => {
   res.send("Hello, world!");
